Prefer shortest match if multiple matches found at same offset.
Previously we would prefer to match the longest possible match. This is
not the semantics expected e.g., in spicy, see zeek/spicy#1114.

With this patch we now prefer the shortest match if multiple matches
where same at the same offset, given that they are all in at the
left-most offset.