Add variant of `jrx_regset_add` that allows to provide a group match ID for the new pattern.
This reworks the API to provide group match IDs when compiling
multiple regular expressions jointly for parallel matching.
Previously, one had to add `{#ID}` to the end of a regular expression
to provide a custom match ID, which was quite a hack. This change
removes the code parsing such `{...}` add-ons,  and instead adds a API
function `jrx_regset_add2` that receives the ID explicitly through an
addtional argument. Internally, we can then also refactor and cleanup
how we pass theses IDs around to where there are needed.

For now, this change breaks providing custom group IDs when running
our testing tool `retest` (because the old parsing code is gone, and
it doesn't yet know how to use `jrx_regset_add2`). It will soon
re-gain that capability as part of a rewrite.

As prep work for subsequent new functionality, this also adds a 2nd
new argument  to `jrx_regset_add2`: `cflags`  allows passing
pattern-specific regexp flags on top of the global flags provided when
initializing the set of regular expressions as a whole. But for now, no
such local flags are actually supported yet.