Speed up CCL matching.
We now keep a lookup table for all CCL codejoints < 256, so that in
the most common case we do not need to iterate through all the ranges
anymore. In addition to the table, we also maintain a separate range
list that records only ranges exceeding the ASCII range.

This seems to speed up the case of a single codejoint slightly as
well.